+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Bug Fixes & Changes
 
 - Added a "New Frosted Ice Cracking" config option to control Frosted Ice's updated cracking (playing sounds and spawning particles)
 - Added a config option for whether or not Dripleaves will use updated Redstone powering functionality
+- Swapped the placement of the Misc and Worldgen config tabs
 
 - Removed the Fabric Loader version requirement in hopes it will fix issues with Quilt
 - Changed the grammar of Wilder Wild's advancements to align with Vanilla's